+++ CHANGELOG.md
@@ -832,3 +832,4 @@ https://github.com/haskell/cabal/issues/6481 for more information.
     - guards, although cool, replaced w/ pattern matching
     - very close. appears to work for all test input
     - missing 2 digit numbers
+    - proved it works on test input, To-Do: parse 2 digit numbers